# Rework statetree to be a tree