--- a/Cargo.toml
+++ b/Cargo.toml
@@ -14,5 +14,4 @@ members = [
 ]
 
 [profile.release]
-# lto = true
 debug = true