This tries to make the `StateTree` an actual tree shape, rather than a map of objectids to values.

This means we can then treat it as mutable when applying diffs and other things. This in turn means less small intermediate objects are created which thrash the allocator.

This superseeds https://github.com/automerge/automerge-rs/pull/87

Not all the tests are passing yet but please take a look and let me know if things look off or you think things should be changed :)