@orionz 2021-06-01
Nice?  What kind of impact on performance will this have?