@jeffa5 2021-06-01
From running the `perf` binary in this branch: (note the maps scenario has 10x fewer iterations in the old)

old
```
maps x1000 9.249856094s
seqs x10000 17.712264988s
rand x10000 5.622552404s

processed 1000 changes in 12.098478741s
processed 2000 changes in 26.658259932s
processed 3000 changes in 49.485821416s
processed 4000 changes in 88.180449921s
processed 5000 changes in 133.924567291s
```

new
```
maps x10000 751.935071ms 
seqs x10000 215.179621ms
rand x10000 9.648037191s 

processed 1000 changes in 1.915281184s
processed 2000 changes in 3.838819054s
processed 3000 changes in 6.540369328s
processed 4000 changes in 10.681297088s
processed 5000 changes in 15.389896765s
```