@jeffa5 2021-06-03
This is now at a point where most of the time (from looking at a flamegraph) is spent doing tree rebalancing in the `Vector`. To get further we'll likely need to do some stuff like rle encoding (https://github.com/josephg/text-crdt-rust) or using indices for actors to have smaller items (or both!).

The only failing tests seem to be related to cursors. I've not been able to find an implementation of cursors in the js side code and am not too sure on their intended semantics. If anyone else can help out with that then please do!