@jeffa5 2021-06-09
@alexjg Yes, cursors still need to be made to work again, do you know if they are implemented in the js side?