@ept 2021-06-09
@jeffa5 The JS implementation of cursors is currently languishing on an unmerged PR automerge/automerge#313. It needs a bit more work before it's finished and my priorities are currently elsewhere, so I think it would be fine to leave cursors aside for now and come back to them later.