@jeffa5 2021-06-10
@alexjg I think this is ready for another look now. I'd like to leave cursors alone for now (currently not working) and fix them up separately.

The new `check_diff` bits should be covering most things that can generate errors, again some may need fixing up in the future but since an error would previously leave the user with no state I think this is still an improvement.