@alexjg 2021-06-13
I have some thoughts on how to make the `unreachable` calls in `resolved_path.rs` a bit safer but I think that can be handled later. Other than that I think this is a go for merge!